QPGreen

Documentation for QPGreen.

QPGreen.IntegrationCache

QPGreen.IntegrationParameters

QPGreen.S_even

QPGreen.S_odd

QPGreen.S₀

QPGreen.Yε

QPGreen.Yε_1st_der

QPGreen.Yε_2nd_der

QPGreen.asin_ls

QPGreen.bernoulli

QPGreen.eigfunc_expansion

QPGreen.eigfunc_expansion_derivative

QPGreen.ewald

QPGreen.fm_method_calculation

QPGreen.fm_method_preparation

QPGreen.fm_method_preparation_derivative

QPGreen.f₁

QPGreen.f₂

QPGreen.get_F̂ⱼ

QPGreen.get_K̂ⱼ!

QPGreen.get_t

QPGreen.get_Ĥⱼ

QPGreen.h₁

QPGreen.h₂

QPGreen.image_expansion

QPGreen.lattice_sums_calculation

QPGreen.lattice_sums_preparation

QPGreen.rfftshift_normalization!

QPGreen.χ

QPGreen.χ_der

QPGreen.IntegrationCache — Type

struct IntegrationCache{T1<:Real, T2<:Signed} <: QPGreen.AbstractIntegrationCache

Structure storing the normalization factor and the parameters of integration for the cutoff functions.

normalization::Real: Normalization factor

params::QPGreen.IntegrationParameters: Parameters of integration

QPGreen.IntegrationParameters — Type

struct IntegrationParameters{T1<:Real, T2<:Signed}

Structure storing the parameters of integration for the cutoff functions.

a::Real: Lower bound

b::Real: Upper bound

order::Signed: Order of the cutoff function

QPGreen.S_even — Method

S_even(l, β, k, d, M)

Compute the lattice sum Sₗ for even values of l. Input arguments:

l: integer value.

β: parameter β.

k: parameter k.

d: parameter d.

M: number of terms in the sum.

Returns the value of the lattice sum Sₗ for even values of l.

QPGreen.S_odd — Method

S_odd(l, β, k, d, M)

Compute the lattice sum Sₗ for odd values of l.

Input arguments:

l: integer value.

β: parameter β.

k: parameter k.

d: parameter d.

M: number of terms in the sum.

Returns the value of the lattice sum Sₗ for odd values of l.

QPGreen.S₀ — Method

S₀(β, k, d, M)

Compute the lattice sum S₀.

Input arguments:

β: parameter β.

k: parameter k.

d: parameter d.

M: number of terms in the sum.

Returns the value of the lattice sum S₀.

QPGreen.Yε — Method

Yε(x, cache)

Build the cutoff function Yε. Input arguments:

x: point at which the cutoff function is evaluated

cache: see IntegrationCache

Returns the value of the cutoff function at x.

QPGreen.Yε_1st_der — Method

Yε_1st_der(x, cache)

Build the derivative of the cutoff function Yε. Input arguments:

x: point at which the derivative of the cutoff function is evaluated

cache: see IntegrationCache

Returns the value of the derivative of the cutoff function at x.

QPGreen.Yε_2nd_der — Method

Yε_2nd_der(x, cache)

Build the 2nd derivative of the cutoff function Yε. Input arguments:

x: point at which the 2nd derivative of the cutoff function is evaluated

cache: see IntegrationCache

Returns the value of the 2nd derivative of the cutoff function at x.

QPGreen.asin_ls — Method

asin_ls(x)

Compute the inverse sine function (arcsine function) for values outside of the interval [-1,1]. Input arguments:

x: real values outside the interval [-1,1].

Returns the inverse sine function (arcsine function) evaluated at x. For values inside the interval [-1,1], use the standard asin function.

QPGreen.bernoulli — Method

bernoulli(n, x)

Compute the Bernoulli polynomial of degree n evaluated at x. Input arguments:

n: degree of the Bernoulli polynomial.

x: value at which the polynomial is evaluated.

Returns the nth Bernoulli polynomial evaluated at x.

QPGreen.eigfunc_expansion — Method

eigfunc_expansion(z, csts; period=2π, nb_terms=100)

z: coordinates of the difference between the target point and source point

csts: Named tuple of the constants for the problem definition

Returns the value of the α-quasi-periodic Green function for 2D Helmholtz equation at the point z defined by the basic eigenfunction expansion.

QPGreen.eigfunc_expansion_derivative — Method

eigfunc_expansion_derivative(z, csts; period=2π, nb_terms=100)

Calculate the derivative of the Green's function using the eigenfunction expansion. Input arguments:

z: 2D point at which the derivative is evaluated

csts: Named tuple of the constants for the problem definition

Keyword arguments:

period: period of the Green's function

nb_terms: number of terms in the series expansion

Returns the value of the derivative of the Green's function.

QPGreen.ewald — Method

ewald(z, csts)

Evaluation of the Green's function using Ewald's method. Input arguments:

z: coordinates.

csts: tuple of constants (a, M₁, M₂, N, β, k, d).

Returns the value of the Green's function at the point z.

QPGreen.fm_method_calculation — Method

fm_method_calculation(x, csts, interp_cubic, cache_Yε; nb_terms=100)

Calculation step of the FFT-based algorithm. Input arguments:

x: given as a 2D array

csts: tuple of the constants (α, k, c, c̃, ε, order)

interp_cubic: bicubic interpolation function

cache_Yε: cache for the cut-off function Yε

Keyword arguments:

  - nb_terms: number of terms in the series expansion

Returns the approximate value of the Green's function G(x).

QPGreen.fm_method_preparation — Method

fm_method_preparation(csts, grid_size)

Preparation step of the FFT-based algorithm. Input arguments:

csts: tuple of constants (α, k, c, c̃, ε, order) (recommended value: 0.4341)

grid_size: size of the grid

Returns the Fourier coefficients of the function Lₙ.

QPGreen.fm_method_preparation_derivative — Method

fm_method_preparation_derivative(csts, grid_size)

Preparation of the Fourier coefficients for the first order derivative of the Green's function.

QPGreen.f₁ — Method

f₁(x, Yε)

Calculate the function f₁. Input arguments:

x: point at which the function is evaluated

Yε: cut-off function Yε

Returns the value of the function f₁.

QPGreen.f₂ — Method

f₂(x, cache)

Calculate the function f₂. Input arguments:

x: point at which the function is evaluated

cache: cache for the cut-off function Yε

Returns the value of the function f₂.

QPGreen.get_F̂ⱼ — Method

get_F̂ⱼ(j₁, j₂, c̃, Φ̂₁ⱼ, Φ̂₂ⱼ, F̂₁ⱼ₀, type_α)

Calculate the Fourier coefficients F̂ⱼ. Input arguments:

j₁: first index

j₂: second index

c̃: parameter of the periodicity

Φ̂₁ⱼ: Fourier coefficients of the function Φ₁

Φ̂₂ⱼ: Fourier coefficients of the function Φ₂

F̂₁ⱼ₀: Fourier coefficient of the function F₁ at |j| = 0

type_α: type of the parameter α

Returns the Fourier coefficients F̂ⱼ.

QPGreen.get_K̂ⱼ! — Method

get_K̂ⱼ!((K̂ⱼ, csts, N, i, fft_cache, cache, p)

Mutating function that computes the Fourier coefficients K̂ⱼ. Input arguments:

K̂ⱼ: matrix to store the Fourier coefficients

csts: Named tuple of the constants for the problem definition

N: size of the grid

i: index of the grid points

fft_cache: cache for the FFT

cache: cache for the cut-off function Yε

p: plan for the FFT

Returns the Fourier coefficients K̂ⱼ.

QPGreen.get_t — Method

get_t(x)

Find the value of t in the interval [-π, π[ such that x = 2nπ + t. Input arguments:

x: point at which the function is evaluated

Returns the value of t.

QPGreen.get_Ĥⱼ — Method

get_Ĥⱼ(j₁, j₂, csts, F̂₁ⱼ, F̂₂ⱼ, ρ̂ₓⱼ, Φ̂₃ⱼ, ĥ₁ⱼ, ĥ₂ⱼ, Ĥ₁ⱼ₀, type_α)

Calculate the Fourier coefficients Ĥⱼ. Input arguments:

- j₁: first index
- j₂: second index
- csts: Named tuple of the constants for the problem definition
- F̂₁ⱼ: Fourier coefficients of the function F₁
- F̂₂ⱼ: Fourier coefficients of the function F₂
- ρ̂ₓⱼ: Fourier coefficients of the function ρₓ
- Φ̂₃ⱼ: Fourier coefficients of the function Φ₃
- ĥ₁ⱼ: Fourier coefficients of the function h₁_reduced
- ĥ₂ⱼ: Fourier coefficients of the function h₂_reduced
- Ĥ₁ⱼ₀: Fourier coefficient of the function H₁ at |j| = 0
- type_α: type of the parameter α

Returns the Fourier coefficients Ĥⱼ.

QPGreen.h₁ — Method

h₁(x, csts, cache)

Calculate the function h₁. Input arguments:

x: point at which the function is evaluated

csts: Named tuple of the constants for the problem definition

Returns the value of the function h₁.

QPGreen.h₂ — Method

h₂(x, csts, cache)

Calculate the function h₂. Input arguments:

x: point at which the function is evaluated

csts: Named tuple of the constants for the problem definition

Returns the value of the function h₂.

QPGreen.image_expansion — Method

image_expansion(z, csts; period=2π, nb_terms=100)

z: coordinates of the difference between the target point and source point

csts: Named tuple of the constants for the problem definition

Returns the value of the α-quasi-periodic Green function for 2D Helmholtz equation at the point z defined by the basic image expansion.

QPGreen.lattice_sums_calculation — Method

lattice_sums_calculation(x, csts, Sₗ; c=0.6, nb_terms=100)

Compute the Green's function using lattice sums. Input arguments:

x: evaluation point

csts: tuple of constants (β, k, d, M, L)

Sₗ: lattice sum coefficients

c: cutoff value

nb_terms: number of terms in the sum

Keyword arguments:

c: cutoff value

nb_terms: number of terms in the sum

Returns the value of the Green's function at the point x.

QPGreen.lattice_sums_preparation — Method

lattice_sums_preparation(csts)

Perform the preparation step for the lattice sums algorithm. Input arguments:

csts: tuple of constants (β, k, d, M, L) representing quasi-periodicty parameter, wave number , period, number of terms, and number of terms in the sum respectively.

Returns the lattice sum coefficients.

QPGreen.rfftshift_normalization! — Method

rfftshift_normalization!(Φ̂₁ⱼ, fft_Φ₁_eval, N, c̃)

Shift the Fourier coefficients obtained by rfft and normalize them.

QPGreen.χ — Method

χ(x, cache)

Build the cutoff function χ. Input arguments:

x: point at which the cutoff function is evaluated

cache: see IntegrationCache

Returns the value of the cutoff function at x.

QPGreen.χ_der — Method

χ_der(x, cache)

Build the derivative of the cutoff function χ. Input arguments:

x: point at which the derivative of the cutoff function is evaluated

cache: see IntegrationCache

Returns the value of the derivative of the cutoff function at x.